+++ client/api_error.go
@@ -6,9 +6,7 @@
 package toxiproxy
 
 import (
-	"encoding/json"
 	"fmt"
-	"net/http"
 )
 
 type ApiError struct {
@@ -18,17 +16,4 @@ type ApiError struct {
 
 func (err *ApiError) Error() string {
 	return fmt.Sprintf("HTTP %d: %s", err.Status, err.Message)
-}
-
-func checkError(resp *http.Response, expectedCode int, caller string) error {
-	if resp.StatusCode != expectedCode {
-		apiError := new(ApiError)
-		err := json.NewDecoder(resp.Body).Decode(apiError)
-		if err != nil {
-			apiError.Message = fmt.Sprintf("Unexpected response code, expected %d", expectedCode)
-			apiError.Status = resp.StatusCode
-		}
-		return fmt.Errorf("%s: %v", caller, apiError)
-	}
-	return nil
 }